Refactor skin handling in character commands
- Remove unused `skinPartBuilder` method from `SkinFactory` and `PaperSkinFactory`
- Update skin parts handling in `EntityCodecs` and `CharacterSkinLayerCommand`
- Rename `SkinLayer` class to `plugin.model` package
