+++ api/src/main/java/net/thenextlvl/character/skin/SkinFactory.java
@@ -1,7 +1,6 @@
 package net.thenextlvl.character.skin;
 
 import com.destroystokyo.paper.profile.ProfileProperty;
-import org.jetbrains.annotations.Contract;
 
 import java.io.File;
 import java.net.URL;
@@ -11,7 +10,4 @@ public interface SkinFactory {
     CompletableFuture<ProfileProperty> skinFromFile(File image, boolean slim);
 
     CompletableFuture<ProfileProperty> skinFromURL(URL url, boolean slim);
-
-    @Contract(value = " -> new", pure = true)
-    SkinPartBuilder skinPartBuilder();
 }
