+++ src/main/java/net/thenextlvl/character/plugin/model/SkinLayer.java
@@ -1,4 +1,4 @@
-package net.thenextlvl.character.skin;
+package net.thenextlvl.character.plugin.model;
 
 import net.kyori.adventure.translation.Translatable;
 import org.jetbrains.annotations.Contract;
